requirement.txt
channels==4.3.1
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
daphne==4.2.1
redis==6.4.0
scapy==2.6.1






